re-organize the `cosmos` package
